requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
filelock==3.18.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
griffe==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
llama-index-core==0.12.28
llama-index-embeddings-huggingface==0.5.2
llama-index-llms-huggingface-api==0.4.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.3.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.50.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
